Use ReadBytes when reading ASCII characters
It is wasteful to require the framework to convert all the bytes to
chars because the buffer is usually larger than the resulting a string
and we only use the data up to the first NUL byte.

Renamed the StringFromChar method to StringFromASCIIBytes and changed
its parameter to a byte array.
Updated the GetFilterDataFromParmBytes method to use ReadBytes instead
of ReadChars.