Support the \r escape sequence in AFS source code
Updated the ReadAfsString method to stop reading the line when it
encounters the \r escape sequence.
This simplifies the code and allows any white space that may follow an
escape sequence to be trimmed after the next line is read.

Changed the ReadAfsString method to return null when it encounters a
blank line.
This allows the LoadAfs method to distinguish the lines that separate
the source code for the different channels from a line that contained
only white space.