Save the filter defaults in the effect token
Added a ResetData property to the effect token.
This allows the default control values of a loaded filter to be saved
in the effect token instead of loading the filter from disk to retrieve
the default values.

Removed the code that reloads the last filter and the unused second
parameter from the GetFilterfromFFLOffset, LoadFFL and LoadFilter
methods.
Removed the ResetTokenDataInfo method.