Cache the FFL filter data in the TreeNode tag
Changed the FFLoadSave.LoadFFL method to return a list of filters
instead of a the list of TreeNodes.
Updated the PdnFFConfigDialog.LoadFFL method to create the TreeNodes
from the filter list.

Changed the TreeNode creation code to set the filter data as the node
tag and updated the FFLtreeView1_AfterSelect method to use the filter
data in the tag.
Removed the GetFilterfromFFLOffset method.